Extracted ng2 specific providers from classes
