+++ lib/index.ts
@@ -1,5 +1,5 @@
 export { Effect } from './metadata';
 export { mergeEffects } from './effects';
-export { runEffects } from './run-effects';
+export { runEffects } from './ng2';
 export { StateUpdate, StateUpdates } from './state-updates';
 export { toPayload, all } from './util';
