+++ lib/state-updates.ts
@@ -1,18 +1,14 @@
 import { withLatestFrom } from 'rxjs/operator/withLatestFrom';
-import { map } from 'rxjs/operator/map';
 import { filter } from 'rxjs/operator/filter';
 import { Observable } from 'rxjs/Observable';
 import { ReplaySubject } from 'rxjs/ReplaySubject';
 import { Action, Dispatcher, State } from '@ngrx/store';
-import { Injectable, Provider } from '@angular/core';
-
 
 export interface StateUpdate<S> {
   state: S;
   action: Action;
 }
 
-@Injectable()
 export class StateUpdates<S> extends ReplaySubject<StateUpdate<S>> {
   constructor(actions$: Dispatcher, state$: State<S>) {
     super(1);
@@ -34,7 +30,3 @@ export class StateUpdates<S> extends ReplaySubject<StateUpdate<S>> {
     );
   }
 }
-
-export const STATE_UPDATES_PROVIDER = new Provider(StateUpdates, {
-  useClass: StateUpdates
-});
