+++ lib/util.ts
@@ -1,4 +1,3 @@
-import { Action } from '@ngrx/store';
 import { StateUpdate } from './state-updates';
